Move serialization code for SLB_level_conv to its own namespace
This is part of a cleanup to add classes for all SLB
files, this IO code is anyway going to go away with
the new "generic serializers"
