+++ SAGESharp/SAGESharp.csproj
@@ -132,14 +132,14 @@
     <Compile Include="SLB\IO\ISLBBinaryWriter.cs" />
     <Compile Include="SLB\IO\StringBinaryWriter.cs" />
     <Compile Include="SLB\Level\Conversation\Character.cs" />
-    <Compile Include="SLB\Level\Conversation\CharacterBinaryWriter.cs" />
-    <Compile Include="SLB\Level\Conversation\ConversationBinaryAccessor.cs" />
-    <Compile Include="SLB\Level\Conversation\ConversationBinaryWriter.cs" />
-    <Compile Include="SLB\Level\Conversation\ConversationFooterGenerator.cs" />
     <Compile Include="SLB\Level\Conversation\Frame.cs" />
-    <Compile Include="SLB\Level\Conversation\FrameBinaryWriter.cs" />
     <Compile Include="SLB\Level\Conversation\Info.cs" />
-    <Compile Include="SLB\Level\Conversation\InfoBinaryWriter.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\CharacterBinaryWriter.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryAccessor.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryWriter.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\ConversationFooterGenerator.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\FrameBinaryWriter.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\InfoBinaryWriter.cs" />
     <Compile Include="SLB\Level\Conversation\LineSide.cs" />
     <Compile Include="Utils\ASCIIExtensions.cs" />
     <Compile Include="Utils\NumericExtensions.cs" />
