+++ SAGESharpTests/SAGESharpTests.csproj
@@ -95,16 +95,16 @@
     <Compile Include="SLB\IO\FooterEntryTests.cs" />
     <Compile Include="SLB\IO\SLBFooterWriterTests.cs" />
     <Compile Include="SLB\IO\StringBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\CharacterBinaryWriterTests.cs" />
     <Compile Include="SLB\Level\Conversation\CharacterEqualityTests.cs" />
-    <Compile Include="SLB\Level\Conversation\ConversationBinaryAccessorTests.cs" />
-    <Compile Include="SLB\Level\Conversation\ConversationBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\ConversationFooterGeneratorTests.cs" />
     <Compile Include="SLB\Level\Conversation\Defaults.cs" />
-    <Compile Include="SLB\Level\Conversation\FrameBinaryWriterTests.cs" />
     <Compile Include="SLB\Level\Conversation\FrameEqualityTests.cs" />
-    <Compile Include="SLB\Level\Conversation\InfoBinaryWriterTests.cs" />
     <Compile Include="SLB\Level\Conversation\InfoEqualityTests.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\CharacterBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryAccessorTests.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\ConversationFooterGeneratorTests.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\FrameBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\Conversation\IO\InfoBinaryWriterTests.cs" />
     <Compile Include="SLB\Level\Conversation\SerializationTests.cs" />
     <Compile Include="SLB\Level\Conversation\TestData.cs" />
     <Compile Include="SLB\Point3DEqualityTests.cs" />
