+++ SAGESharpTests/SLB/Level/Conversation/IO/ConversationBinaryAccessorTests.cs
@@ -9,7 +9,7 @@ using SAGESharp.Testing;
 using System.Collections.Generic;
 using System.IO;
 
-namespace SAGESharp.SLB.Level.Conversation
+namespace SAGESharp.SLB.Level.Conversation.IO
 {
     class ConversationBinaryAccessorTests
     {
