+++ SAGESharpTests/SLB/Level/Conversation/IO/ConversationBinaryWriterTests.cs
@@ -11,7 +11,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace SAGESharp.SLB.Level.Conversation
+namespace SAGESharp.SLB.Level.Conversation.IO
 {
     class ConversationBinaryWriterTests
     {
